refactor(inspect): Changed reliance `__code__` to replace it with `inspect.signature()`.
Complete Fork: Name change Included and coming interface changes.

To be able to create simulations quickly this project will a large refactor. Using a lot of the modern design patterns to allow for easier simulation formation. To simply the development it'll be named differently, with the option to let the blockscience team adopt it as their own.

### Problem Context:
---
Here's the context of the issue. There was an issue in using decorators around things like the state and policy functions. In order to allow for that I needed to change a piece of code that checks for the `__code__` variable.  It was replaced with

BREAKING CHANGE:  named to `prima` instead of cadcad.